% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_extract.R
\name{clim_extract}
\alias{clim_extract}
\title{Acquire climate data for climatic diagram}
\usage{
clim_extract(file)
}
\arguments{
\item{file}{A data.frame(see details in dataset "locdata") with the following 5 columns:
    \itemize{
    \item \code{No}: Serial number of the locations
    \item \code{location}: Abbreviation of the locations
    \item \code{lon}: Longitude of the locations in decimal digit
(East longitude are represented by negative numbers)
    \item \code{lat}: Latitude of the locations in decimal digit
(South latitude are represented by negative numbers)
    \item \code{altitude}: Altitude of the locations
    }
    Other columns with information is allowed behind the columns above
    but would be discarded in following process.}
}
\value{
A data.frame with annual average precipitation, annual average minimum
temperature and annual average max temperature of different months, as well as
other essential information of every location.
\itemize{
   \item \code{No,location,lon,lat}: information of the station,the the same as
which in parameter code{file}
    \item \code{type}: Labels of the climate data among precipitation and temperature
    \item \code{1-12}: Values of specific type of climate data,the names of columns
represent monthly values from January to December
}
}
\description{
\code{clim_extract} acquires climate data which are esstential for
Walter & Lieth climatic diagram plotting according to the provided coordinate of
the location.
}
\details{
The function extracts precipitation and temperature from arranged
Worldclim Historical monthly weather data(https://worldclim.org/data/monthlywth.html)
and arranges them to a data frame for plotting the climatic diagram.
}
\examples{
{
#import data of stations
data("locdata")
a <- "D:/climplot/mean_mintemp"
b <- "D:/climplot/mean_maxtemp"
c <- "D:/climplot/mean_prec"
#extraction of climate data
cli <- clim_extract(locdata,a,b,c)
}
}
\references{
{Guijarro J A (2023). climatol: Climate Tools
(Series Homogenization and Derived Products), 4.0.0.,
https://CRAN.R-project.org/package=climatol

Walter H & Lieth H (1960): Klimadiagramm Weltatlas. G. Fischer, Jena.}
}
\keyword{internal}
