% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_extract.R
\name{clim_extract}
\alias{clim_extract}
\title{Obtain climate data for climatic diagram drawing online}
\usage{
clim_extract(file, path = tempdir())
}
\arguments{
\item{file}{A data.frame(see details in dataset \code{locdata})
    with the following 5 columns:
    \itemize{
    \item \code{No}: Serial number of the locations
    \item \code{location}: Abbreviation of the locations
    \item \code{lon}: Longitude of the locations in decimal digit
(West longitude are represented by negative numbers)
    \item \code{lat}: Latitude of the locations in decimal digit
(South latitude are represented by negative numbers)
    \item \code{altitude}: Altitude of the locations
    }
    Other columns with information is allowed behind the columns above
    but would be discarded in following process.}

\item{path}{character.A Path for storing the downloaded data,
avoid downloading the same data many times over and
guards against service interruptions. Deafault is
 tempdir().}
}
\value{
A data.frame with annual average precipitation, annual average
temperature, annual average minimum temperature and
annual average max temperature of 12 months, as well as
other essential information of every location.
Note that some locations don't have data of annual average temperature,
so they would be replace by values averaged by
annual average minimum temperature and annual average maximum temperature.


\itemize{
   \item \code{No,location,lon,lat}: Information of the locations,
   the the same as which in parameter \code{file}.
    \item \code{type}: The labels of climate data,
    encompassing annual average precipitation,
    annual average minimum temperature,annual average temperature and
    annual average maximum temperature.
    \item \code{1-12}: The column names of the particular climate data type
    correspond to monthly values ranging from January to December.
}
See more details in dataset \code{plotdata}.
}
\description{
\code{clim_extract} acquires crucial climate data from online
datasets for drawing Walter & Lieth climatic diagrams
based on the provided location coordinates.
}
\details{
This function extracts precipitation and temperature from
Worldclim Historical monthly weather
data in 2010-2019
(Version of 2.5 minutes,https://worldclim.org/data/monthlywth.html),
containing annual average precipitation, annual average
temperature,annual average minimum
temperature and annual average max temperature of 12 months. It downloads the
climate data, extracts and arranges them to a data frame for drawing
the climatic diagram.
}
\examples{
{
#import data of locations
x <- data.frame(No = "1", location = "test",
lon = 0, lat = -30, altitude = 20)
y <- clim_extract(x)
}

}
