% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_plot.R
\name{clim_plot}
\alias{clim_plot}
\title{Walter & Lieth climatic diagram drawing}
\usage{
clim_plot(data, mlab = "", pcol = "blue", tcol = "red",
wcol = "green", dcol = "orange",
pfcol = "#79e6e8", sfcol = "#09a0d1", ylabel = FALSE, ylab1 = NA,
ylab2 = NA, xlab = "Month",
showfrost = FALSE, shem = FALSE, line_p3 = FALSE, line_p50 = FALSE,
temp_extreme = FALSE, margen = c(4, 4, 5, 4),per = NA)
}
\arguments{
\item{data}{A data.frame(see details in the dataset \code{plotdata})
with the following 18 columns:
    \itemize{
   \item \code{No}: Serial number of the locations.
    \item \code{location}: Abbreviation of the locations.
    \item \code{lon}: Longitude of the locations in decimal digit
    (negative numbers indicating west longitude).
    \item \code{lat}: Latitude of the locations in decimal digit
    (negative numbers indicating south latitude).
    \item \code{altitude}: Altitude of the locations.
    \item \code{type}: The labels of climate data, encompassing
    annual average precipitation,
    annual average minimum temperature and
    annual average maximum temperature.
    In the event that Frost=True, it will also incorporate
    annual extreme minimum temperature.
    \item \code{1-12}: The column names of the particular climate data type
    correspond to monthly values ranging from January to December.
}
    Other columns containing information may be placed
    after the aforementioned columns, but will be disregarded
    in subsequent processes.}

\item{mlab}{A vector of 12 monthly labels for the X axis.Deafult is numeric.
Alternatively, set as 'en' or 'es' can be specified to display
the month names in English or Spanish respectively.
(original parameter of the referenced function \code{diagwl()}).}

\item{pcol}{Color of the precipitation curve. Default is "blue".}

\item{tcol}{Color of the temperature curve. Default is "red".}

\item{wcol}{Color of the humid season polygon. Default is "green".}

\item{dcol}{Color of the arid season polygon. Default is "orange".}

\item{pfcol}{Color of the potential frosty months. Default is "#79e6e8".
No use when showfrost= FALSE.}

\item{sfcol}{Color of the confirmed frosty months. Default is "#09a0d1".
No use when showfrost= FALSE.}

\item{ylabel}{A logical value for whether using customized label of y axis.
Default is FALSE.}

\item{ylab1}{A character value for label of y axis of temperature.
Default is NA.}

\item{ylab2}{A character value for label of y axis of precipitation.
Default is NA.}

\item{xlab}{A character value for label of x axis. Default is "Month".}

\item{showfrost}{A logical value for whether marking forst months.
Default is FALSE.}

\item{shem}{A logical value for whether keeping the summer period in
the central zone of the graphic
(the diagram will begin the plot with the July data). Default is FALSE.
Useful when the location is in southern hemisphere
(original parameter of the referenced function \code{diagwl()}).}

\item{line_p3}{A logical value for whether displaying
a supplementary precipitation line referenced to three times the temperature.
Default is FALSE
(original parameter of the referenced function \code{diagwl()}).}

\item{line_p50}{A logical value for whether displaying a supplementary
precipitation line in 50°C-100mm. Default is FALSE.}

\item{temp_extreme}{A logical value for whether displaying
the extreme temperature. Default is FALSE.}

\item{margen}{A vector to control the range of plot.
Default is "c(4, 4, 5, 4)"}

\item{per}{a parameter from the referenced function \code{diagwl()}.
No use temporarily.}
}
\value{
A Walter & Lieth climatic diagram of the provided locations,
including annual variation
of temperature and precipitation ,
as well as the time of humid and arid season.
}
\description{
\code{clim_plot} plots the Walter & Lieth climatic diagram
with the climate data of different locations.
It is based on the method of function
\code{diagwl()}from package \code{climatol},
but offers additional parameters for color selection, axis labeling,
and location information display
to meet diverse research and publishing needs.
}
\details{
The function extracts precipitation and temperature from arranged
Worldclim Historical monthly weather data
(https://worldclim.org/data/monthlywth.html)
and arranges them to a data.frame for drawing the climatic diagram.
}
\examples{
{
  data("plotdata")
  test <- subset(plotdata, No == 10)
  clim_plot(
    data = test, ylabel = TRUE,
    ylab1 = "Temperature(\U{00B0}C)",
    ylab2 = "Precipitation(mm)",
    line_p50 = TRUE
  )

  # use loop to plot multiple diagrams simultaneously
  list <- unique(plotdata$No)
  par(mfrow = c(1, 1))
  for (i in 1:5) {
    k <- list[i]
    sub <- subset(plotdata, No == k)
    clim_plot(
      data = sub, ylabel = TRUE,
      ylab1 = "Temperature(\U{00B0}C)",
      ylab2 = "Precipitation(mm)",
      line_p50 = TRUE
    )
  }
}

}
