% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locdata}
\alias{locdata}
\title{An example of data format for "clim_extract"}
\format{
A data frame with 10 rows and 6 variables:
\describe{
    \item{No}{Serial number of the locations}
    \item{location}{Abbreviation of the locations}
    \item{lon}{Longitude of the locations in decimal digit}
    \item{lat}{Latitude of the locations in decimal digit}
    \item{altitude}{Altitude of the locations}
    \item{name}{Full name of the locations}
    ...
    }
}
\source{
\url{NULL}
}
\description{
A dataset containing the abbreviation, coordinates and altitude
of different locations.
}
\details{
5 columns of this dataset: "No", "location", "lon", "lat", "altitude"
are essential for plotting, the columns "name" just a remark.
}
\keyword{internal}
