% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_plot.R
\name{clim_plot}
\alias{clim_plot}
\title{Climatic diagram plotting in a more convenient and customized way}
\usage{
clim_plot(data,mlab="",pcol="blue",tcol="red",wcol="green",dcol="orange",
pfcol="#79e6e8",sfcol="#09a0d1",ylabel=FALSE,ylab1=NA,ylab2=NA,xlab="Month",
ShowForst=FALSE,shem=FALSE,p3line=FALSE,p50line=FALSE,extremeT=FALSE,
margen=c(4,4,5,4),per=NA,...)
}
\arguments{
\item{data}{A data.frame(see details in the dataset \code{plotdata})
with the following 18 columns:
    \itemize{
   \item \code{No}: Serial number of the locations.
    \item \code{location}: Abbreviation of the locations.
    \item \code{lon}: Longitude of the locations in decimal digit
(East longitude are represented by negative numbers).
    \item \code{lat}: Latitude of the locations in decimal digit
(South latitude are represented by negative numbers).
    \item \code{altitude}: Altitude of the locations.
    \item \code{type}: Labels of the climate data,including annual average
    precipitation,annual average minimum temperature and annual average
    maximum temperature.
    \item \code{1-12}: Values of specific type of climate data,the names of columns
represent monthly values from January to December.
}
    Other columns with information is allowed behind the columns above
     but would be discarded in following process.}

\item{mlab}{Vector of 12 monthly labels for the X axis,deafult is numeric,
set to just 'en' or 'es' to use the first letter of
the month names in English or Spanish respectively
(original parameter of the referenced function \code{diagwl()}).}

\item{pcol}{Color of the precipitation curve. Default is "blue".}

\item{tcol}{Color of the temperature curve. Default is "red".}

\item{wcol}{Color of the humid season polygon. Default is "green".}

\item{dcol}{Color of the arid season polygon. Default is "orange".}

\item{pfcol}{Color of the potential frost months. Default is "#79e6e8". No use when
ShowFrost= FALSE}

\item{sfcol}{Color of the accurate frost months. Default is "#09a0d1".
No use when ShowFrost= FALSE}

\item{ylabel}{A logical value for whether using customized label of y axis. Default is FALSE.}

\item{ylab1}{A character value for label of y axis of temperature. Default is NA.}

\item{ylab2}{A character value for label of y axis of precipitation. Default is NA.}

\item{xlab}{A character value for label of x axis. Default is "Month".}

\item{shem}{A logical value for whether keeping the summer period in
the central zone of the graphic
(the diagram will begin the plot with the July data). Default is FALSE. Useful when the
location is in southern hemisphere
(original parameter of the referenced function \code{diagwl()}).}

\item{p3line}{A logical value for whether displaying
a supplementary precipitation line referenced to three times the temperature.
Default is FALSE
(original parameter of the referenced function \code{diagwl()}).}

\item{p50line}{A logical value for whether displaying a supplementary
precipitation line in 50°C-100mm. Default is FALSE.}

\item{extremeT}{A logical value for whether displaying the extreme temperature.
Default is FALSE.}

\item{margen}{A vector to control the range of plot. Default is "c(4,4,5,4)"}

\item{per}{a parameter from the referenced function \code{diagwl()}. No use temporarily.

#'@details The function extracts precipitation and temperature from arranged
Worldclim Historical monthly weather data(https://worldclim.org/data/monthlywth.html)
and arranges them to a data frame for plotting the climatic diagram.}

\item{ShowFrost}{A logical value for whether marking forst months. Default is FALSE.}
}
\value{
A Walter & Lieth climatic diagram of the given location, including the information
of temperature and precipitation period change, as well as the humid and arid season distribution.
}
\description{
\code{clim_plot} plots the Walter & Lieth climatic diagram
with the climate data of different locations. It is based on the method of function
\code{diagwl()}from package \code{climatol}, but release more parameters for color selections,
axis lab setting and information display of the locations, meeting more
requirements of research and publishing.
}
\references{
{Guijarro J A (2023). climatol: Climate Tools
(Series Homogenization and Derived Products), 4.0.0.,
https://CRAN.R-project.org/package=climatol

Walter H & Lieth H (1960): Klimadiagramm Weltatlas. G. Fischer, Jena.}
}
\keyword{internal}
